feat(init): guard database delete with env var
DJANGO_DB_RESET is true by default to maintain existing behavior

+++ bin/init.sh
@@ -1,15 +1,16 @@
 #!/usr/bin/env bash
 set -eux
 
-# construct the path to the database file from environment or default
-
-DB_DIR="${DJANGO_DB_DIR:-.}"
-DB_FILE="${DB_DIR}/django.db"
-
 # remove existing (old) database file
-# -f forces the delete (and avoids an error when the file doesn't exist)
 
-rm -f "${DB_FILE}"
+if [[ ${DJANGO_DB_RESET:-true} = true ]]; then
+    # construct the path to the database file from environment or default
+    DB_DIR="${DJANGO_DB_DIR:-.}"
+    DB_FILE="${DB_DIR}/django.db"
+
+    # -f forces the delete (and avoids an error when the file doesn't exist)
+    rm -f "${DB_FILE}"
+fi
 
 # run database migrations
